Add dedicated Oracle Cloud Infrastructure provider
Separating cloud integrations from Oracle Database improves discoverability and lets OCI users avoid the database provider's dependencies.
