+++ .github/ISSUE_TEMPLATE/1-airflow_bug_report.yml
@@ -179,6 +179,7 @@ body:
         - mongo
         - mysql
         - neo4j
+        - oci
         - odbc
         - openai
         - openfaas
