+++ providers/oci/tests/system/oci/__init__.py
@@ -14,23 +14,3 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-from __future__ import annotations
-
-from unittest import mock
-
-import pytest
-
-from airflow.providers.oracle.hooks.generative_ai import OciGenerativeAIHook
-
-GenerativeAiClient = pytest.importorskip("oci.generative_ai").GenerativeAiClient
-
-
-class TestOciGenerativeAIHook:
-    @mock.patch.object(OciGenerativeAIHook, "get_client", autospec=True)
-    def test_get_conn_creates_and_caches_native_client(self, mock_get_client):
-        hook = OciGenerativeAIHook()
-        client = mock_get_client.return_value
-
-        assert hook.get_conn() is client
-        assert hook.conn is client
-        mock_get_client.assert_called_once_with(hook, GenerativeAiClient)
