+++ providers/oracle/provider.yaml
@@ -19,8 +19,7 @@
 package-name: apache-airflow-providers-oracle
 name: Oracle
 description: |
-    `Oracle Database <https://www.oracle.com/database/technologies/>`__ and
-    `Oracle Cloud Infrastructure <https://www.oracle.com/cloud/>`__ integrations.
+    `Oracle <https://www.oracle.com/database/technologies/>`__
 
 state: ready
 lifecycle: production
@@ -92,14 +91,6 @@ integrations:
       - /docs/apache-airflow-providers-oracle/operators.rst
     logo: /docs/integration-logos/Oracle.png
     tags: [software]
-  - integration-name: Oracle Cloud Infrastructure
-    external-doc-url: https://docs.oracle.com/en-us/iaas/Content/home.htm
-    logo: /docs/integration-logos/Oracle.png
-    tags: [service]
-  - integration-name: OCI Generative AI
-    external-doc-url: https://docs.oracle.com/en-us/iaas/Content/generative-ai/home.htm
-    logo: /docs/integration-logos/Oracle.png
-    tags: [generative-ai, service]
 
 operators:
   - integration-name: Oracle
@@ -125,12 +116,6 @@ hooks:
     python-modules:
       - airflow.providers.oracle.hooks.handlers
       - airflow.providers.oracle.hooks.oracle
-  - integration-name: Oracle Cloud Infrastructure
-    python-modules:
-      - airflow.providers.oracle.hooks.base_oci
-  - integration-name: OCI Generative AI
-    python-modules:
-      - airflow.providers.oracle.hooks.generative_ai
 
 transfers:
   - source-integration-name: Oracle
@@ -141,53 +126,3 @@ connection-types:
   - hook-class-name: airflow.providers.oracle.hooks.oracle.OracleHook
     hook-name: "Oracle"
     connection-type: oracle
-  - hook-class-name: airflow.providers.oracle.hooks.base_oci.OciBaseHook
-    hook-name: "Oracle Cloud Infrastructure"
-    connection-type: oci
-    ui-field-behaviour:
-      hidden-fields:
-        - host
-        - schema
-        - port
-      relabeling:
-        login: User OCID
-        password: Private Key Passphrase
-      placeholders:
-        login: ocid1.user...
-        password: Optional API key passphrase
-        tenancy: ocid1.tenancy...
-        fingerprint: aa:bb:cc:...
-        region: us-chicago-1
-        compartment_id: ocid1.compartment...
-    conn-fields:
-      tenancy:
-        label: Tenancy OCID
-        schema:
-          type:
-            - string
-            - 'null'
-      fingerprint:
-        label: Key Fingerprint
-        schema:
-          type:
-            - string
-            - 'null'
-      private_key_content:
-        label: Private Key Content
-        schema:
-          type:
-            - string
-            - 'null'
-          format: password
-      region:
-        label: Region
-        schema:
-          type:
-            - string
-            - 'null'
-      compartment_id:
-        label: Compartment OCID
-        schema:
-          type:
-            - string
-            - 'null'
