+++ scripts/ci/docker-compose/remove-sources.yml
@@ -94,6 +94,7 @@ services:
       - ../../../empty:/opt/airflow/providers/mongo/src
       - ../../../empty:/opt/airflow/providers/mysql/src
       - ../../../empty:/opt/airflow/providers/neo4j/src
+      - ../../../empty:/opt/airflow/providers/oci/src
       - ../../../empty:/opt/airflow/providers/odbc/src
       - ../../../empty:/opt/airflow/providers/openai/src
       - ../../../empty:/opt/airflow/providers/openfaas/src
