+++ scripts/ci/docker-compose/tests-sources.yml
@@ -107,6 +107,7 @@ services:
       - ../../../providers/mongo/tests:/opt/airflow/providers/mongo/tests
       - ../../../providers/mysql/tests:/opt/airflow/providers/mysql/tests
       - ../../../providers/neo4j/tests:/opt/airflow/providers/neo4j/tests
+      - ../../../providers/oci/tests:/opt/airflow/providers/oci/tests
       - ../../../providers/odbc/tests:/opt/airflow/providers/odbc/tests
       - ../../../providers/openai/tests:/opt/airflow/providers/openai/tests
       - ../../../providers/openfaas/tests:/opt/airflow/providers/openfaas/tests
